requirements.txt
django==3.0.3
djangorestframework==3.11.0
gunicorn==19.9.0
ipython
coverage==5.0.3
django_filter==2.0.0
django-bootstrap4==1.1.1
requests==2.23.0
psycopg2-binary==2.8.4
sw-django-utils==0.0.50
openpyxl==3.0.3
python-dateutil==2.8.1
docxtpl==0.6.3
pillow==7.0.0
python-barcode==0.11.0
sentry-sdk==0.14.3
celery[redis]==4.2.1
redis==3.2.0
